requirements.txt
alabaster==0.7.10
Babel==2.5.1
bleach==2.1.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.9.13
django-appconf==1.0.2
django-bootstrap-breadcrumbs==0.8.2
django-bower==5.2.0
django-compressor==2.2
django-cors-headers==2.1.0
django-filter==1.0.4
django-formtools==1.0
django-guardian==1.4.9
django-mobile==0.7.0
django-recaptcha==1.3.1
django-sortedm2m==1.5.0
django-tastypie==0.14.0
djangorestframework==3.2.5
docutils==0.14
easy-thumbnails==2.4.2
html5lib==1.0b10
icalendar==3.11.7
idna==2.6
imagesize==0.7.1
invoke==0.21.0
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
Pygments==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
rcssmin==1.0.6
reportlab==3.4.0
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.4
sphinxcontrib-websupport==1.0.1
urllib3==1.22
webencodings==0.5.1
